mon: Monitor: rework sync_store_init()
Instead of just initiating a store for a sync, also check the current
store for any on-going sync state and make sure we backup the most
up-to-date monmap we can find, along with the paxos versions from the last
clean state we saw (either from the current store if no sync is on-going,
or from our previous store state if a sync is on-going).

Also make sure we clear out the store before we start the sync; we've also
added a function to make this process the same from other places that
might need to do it, and make sure we compact the store if 'mon compact
on sync' is set to true.

Fixes: #5256 (partially)
Backport: cuttlefish

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>