ceph_argparse.py: change valid() conventions, accept tuple/list args
valid() used to either return False or raise an exception on various
sorts of invalid behavior; this was confusing and unnecessary.  Make
it simply raise exceptions, and adjust callers to cope.

Also, allow validate_one() to accept either a simple string argument
or a tuple/list argument representing "name=val" (for REST API support).
Invalid conditions now include name mismatch with name of arg in
descriptor.  If no value is given, assume value is the same as name.

Signed-off-by: Dan Mick <dan.mick@inktank.com>