osd: clear hb_front if it was previously non-NULL and is now NULL
If we have a real addr for hb_front for a given osd and then a new map
has the osd coming up without an hb_front, we need to clear the addr
field.

Also, improve the debug output in add_heartbeat_peer() so we can tell if
we have no connection or a connection to a blank addr.

Hopefully this is the source of the last round of #5460.

Signed-off-by: Sage Weil <sage@inktank.com>