mon: include any new paxos commits in each sync CHUNK message
We already take note of the paxos version when we begin the sync.  As
sync progresses and there are new paxos commits/txns, include those
and update last_committed, so that when sync completes we will have
a full view of everything that happened during sync.

Note that this does not introduce any compatibility change:

 old provider, new requester:
  - msg->version will no be set except for last msg; no change
 new provider, old requester:
  - provider will send the keys along
  - the requester will not update last_committed until the very end
  - requester will still do a final sync of paxos prefix

Signed-off-by: Sage Weil <sage@inktank.com>